Fix dependencies and upgrade
