+++ apps/climatemappedafrica/package.json
@@ -31,19 +31,19 @@
     "clean": "rm -rf .next .turbo node_modules"
   },
   "dependencies": {
-    "@apollo/client": "^3.10.8",
+    "@apollo/client": "^3.11.1",
     "@commons-ui/core": "^0.1.0",
     "@commons-ui/next": "workspace:*",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
+    "@emotion/react": "^11.13.0",
+    "@emotion/styled": "^11.13.0",
     "@hurumap/core": "workspace:*",
     "@hurumap/next": "workspace:*",
     "@mui/lab": "5.0.0-alpha.155",
-    "@mui/material": "^5.16.1",
-    "@mui/styles": "^5.16.1",
-    "@mui/utils": "^5.16.1",
+    "@mui/material": "^5.16.6",
+    "@mui/styles": "^5.16.6",
+    "@mui/utils": "^5.16.6",
     "@reactour/tour": "^3.7.0",
-    "aws-sdk": "^2.1658.0",
+    "aws-sdk": "^2.1665.0",
     "clsx": "^2.1.1",
     "d3-format": "^3.1.0",
     "deepmerge": "^4.3.1",
@@ -67,20 +67,20 @@
     "swr": "^2.2.5",
     "vega": "^5.30.0",
     "vega-embed": "^6.26.0",
-    "vega-lite": "^5.19.0",
+    "vega-lite": "^5.20.0",
     "vega-tooltip": "^0.33.0",
-    "video.js": "^8.16.1",
+    "video.js": "^8.17.1",
     "videojs-youtube": "^3.0.1",
     "webpack": "^5.93.0",
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
-    "@babel/preset-env": "^7.24.8",
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
     "@commons-ui/testing-library": "workspace:*",
     "@material-ui/codemod": "^4.5.1",
-    "@playwright/test": "^1.45.1",
+    "@playwright/test": "^1.45.3",
     "@storybook/addon-actions": "^7.6.20",
     "@storybook/addon-essentials": "^7.6.20",
     "@storybook/addon-links": "^7.6.20",
@@ -91,7 +91,7 @@
     "@storybook/react": "^7.6.20",
     "@storybook/theming": "^7.6.20",
     "@svgr/webpack": "^8.1.0",
-    "@types/node": "^20.14.10",
+    "@types/node": "^20.14.13",
     "@types/react": "^18.3.3",
     "apollo-link-rest": "^0.9.0",
     "babel-jest": "^29.7.0",
@@ -110,10 +110,10 @@
     "jscodeshift": "^0.15.2",
     "next-sitemap": "^1.9.12",
     "playwright-config-commons-ui": "workspace:*",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "react-test-renderer": "^18.3.1",
     "svg-url-loader": "^8.0.0",
-    "typescript": "^5.5.3"
+    "typescript": "^5.5.4"
   },
   "engines": {
     "node": "20.x"
