+++ apps/codeforafrica/package.json
@@ -33,26 +33,26 @@
     "clean": "rm -rf .next .turbo node_modules"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.614.0",
-    "@aws-sdk/lib-storage": "^3.614.0",
+    "@aws-sdk/client-s3": "^3.621.0",
+    "@aws-sdk/lib-storage": "^3.621.0",
     "@commons-ui/core": "workspace:*",
     "@commons-ui/next": "workspace:*",
-    "@emotion/cache": "^11.11.0",
-    "@emotion/react": "^11.11.4",
+    "@emotion/cache": "^11.13.1",
+    "@emotion/react": "^11.13.0",
     "@emotion/server": "^11.11.0",
-    "@emotion/styled": "^11.11.5",
+    "@emotion/styled": "^11.13.0",
     "@googlemaps/react-wrapper": "^1.1.35",
-    "@mui/material": "^5.16.1",
-    "@mui/utils": "^5.16.1",
+    "@mui/material": "^5.16.6",
+    "@mui/utils": "^5.16.6",
     "@next/env": "^14.2.5",
     "@payloadcms/bundler-webpack": "^1.0.7",
-    "@payloadcms/db-mongodb": "^1.5.2",
+    "@payloadcms/db-mongodb": "^1.7.1",
     "@payloadcms/plugin-cloud-storage": "^1.1.3",
     "@payloadcms/plugin-nested-docs": "^1.0.12",
     "@payloadcms/plugin-sentry": "^0.0.6",
     "@payloadcms/plugin-seo": "^2.3.2",
     "@payloadcms/richtext-slate": "^1.5.2",
-    "@sentry/nextjs": "^8.17.0",
+    "@sentry/nextjs": "^8.21.0",
     "camelcase-keys": "^9.1.3",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
@@ -63,7 +63,7 @@
     "next": "^14.2.5",
     "next-seo": "^6.5.0",
     "nodemailer-sendgrid": "^1.0.3",
-    "payload": "^2.23.1",
+    "payload": "^2.25.0",
     "prop-types": "^15.8.1",
     "qs": "^6.12.3",
     "react": "^18.3.1",
@@ -74,14 +74,14 @@
     "swr": "^2.2.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
+    "@babel/core": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
     "@commons-ui/testing-library": "workspace:*",
-    "@playwright/test": "^1.45.1",
+    "@playwright/test": "^1.45.3",
     "@svgr/webpack": "^8.1.0",
-    "@swc/core": "^1.6.13",
+    "@swc/core": "^1.7.3",
     "@types/express": "^4.17.21",
-    "@types/node": "^20.14.10",
+    "@types/node": "^20.14.13",
     "@types/react": "^18.3.3",
     "babel-jest": "^29.7.0",
     "eslint": "^8.57.0",
@@ -92,10 +92,10 @@
     "jest": "^29.7.0",
     "jest-config-commons-ui": "workspace:*",
     "playwright-config-commons-ui": "workspace:*",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "react-test-renderer": "^18.3.1",
     "ts-node": "^10.9.2",
-    "typescript": "^5.5.3",
+    "typescript": "^5.5.4",
     "webpack": "^5.93.0"
   },
   "engines": {
