+++ apps/promisetracker/package.json
@@ -23,15 +23,15 @@
   "dependencies": {
     "@commons-ui/core": "workspace:*",
     "@commons-ui/next": "workspace:*",
-    "@emotion/react": "^11.11.4",
+    "@emotion/react": "^11.13.0",
     "@emotion/server": "^11.11.0",
-    "@emotion/styled": "^11.11.5",
-    "@mui/icons-material": "^5.16.1",
+    "@emotion/styled": "^11.13.0",
+    "@mui/icons-material": "^5.16.6",
     "@mui/lab": "5.0.0-alpha.155",
-    "@mui/material": "^5.16.1",
-    "@mui/styles": "^5.16.1",
-    "@mui/utils": "^5.16.1",
-    "@sentry/nextjs": "^8.17.0",
+    "@mui/material": "^5.16.6",
+    "@mui/styles": "^5.16.6",
+    "@mui/utils": "^5.16.6",
+    "@sentry/nextjs": "^8.21.0",
     "clsx": "^2.1.1",
     "date-fns": "^2.30.0",
     "form-data": "^4.0.0",
@@ -55,13 +55,13 @@
     "victory": "^36.9.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
-    "@babel/preset-env": "^7.24.8",
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
     "@commons-ui/testing-library": "workspace:*",
-    "@playwright/test": "^1.45.1",
+    "@playwright/test": "^1.45.3",
     "@svgr/webpack": "^8.1.0",
-    "@types/node": "^20.14.10",
+    "@types/node": "^20.14.13",
     "@types/react": "^18.3.3",
     "babel-jest": "^29.7.0",
     "eslint": "^8.57.0",
@@ -72,9 +72,9 @@
     "jest": "^29.7.0",
     "jest-config-commons-ui": "workspace:*",
     "playwright-config-commons-ui": "workspace:*",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "react-test-renderer": "^18.3.1",
-    "typescript": "^5.5.3",
+    "typescript": "^5.5.4",
     "webpack": "^5.93.0"
   },
   "scripts": {
