+++ apps/promisetracker/src/components/Promises/index.js
@@ -94,10 +94,10 @@ function Promises({
         sortedItems = items.sort((a, b) => {
           const aDeadline = a?.promiseDeadline
             ? new Date(a.promiseDeadline).getTime()
-            : a.events?.[0]?.year ?? 0;
+            : (a.events?.[0]?.year ?? 0);
           const bDeadline = b?.promiseDeadline
             ? new Date(b.promiseDeadline).getTime()
-            : b.events?.[0]?.year ?? 0;
+            : (b.events?.[0]?.year ?? 0);
           return bDeadline - aDeadline;
         });
       }
