+++ apps/roboshield/package.json
@@ -17,33 +17,33 @@
   "dependencies": {
     "@commons-ui/core": "workspace:*",
     "@commons-ui/next": "workspace:*",
-    "@emotion/cache": "^11.11.0",
-    "@emotion/react": "^11.11.4",
+    "@emotion/cache": "^11.13.1",
+    "@emotion/react": "^11.13.0",
     "@emotion/server": "^11.11.0",
-    "@emotion/styled": "^11.11.5",
-    "@mui/icons-material": "^5.16.1",
-    "@mui/material": "^5.16.1",
-    "@mui/material-nextjs": "^5.16.1",
-    "@mui/utils": "^5.16.1",
-    "@mui/x-date-pickers": "^7.10.0",
+    "@emotion/styled": "^11.13.0",
+    "@mui/icons-material": "^5.16.6",
+    "@mui/material": "^5.16.6",
+    "@mui/material-nextjs": "^5.16.6",
+    "@mui/utils": "^5.16.6",
+    "@mui/x-date-pickers": "^7.11.1",
     "@next/env": "^14.2.5",
     "@payloadcms/bundler-webpack": "^1.0.7",
-    "@payloadcms/db-mongodb": "^1.5.2",
+    "@payloadcms/db-mongodb": "^1.7.1",
     "@payloadcms/live-preview-react": "^0.2.0",
     "@payloadcms/plugin-cloud-storage": "^1.1.3",
     "@payloadcms/plugin-nested-docs": "^1.0.12",
     "@payloadcms/plugin-sentry": "^0.0.6",
     "@payloadcms/plugin-seo": "^2.3.2",
     "@payloadcms/richtext-slate": "^1.5.2",
-    "@sentry/nextjs": "^8.17.0",
-    "ace-builds": "^1.35.2",
+    "@sentry/nextjs": "^8.21.0",
+    "ace-builds": "^1.35.4",
     "crawler-user-agents": "^1.0.142",
     "date-fns": "^3.6.0",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "next": "^14.2.5",
     "nodemailer-sendgrid": "^1.0.3",
-    "payload": "^2.23.1",
+    "payload": "^2.25.0",
     "react": "^18.3.1",
     "react-ace": "^11.0.1",
     "react-dom": "^18.3.1",
@@ -59,7 +59,7 @@
     "@commons-ui/testing-library": "workspace:*",
     "@svgr/webpack": "^8.1.0",
     "@types/express": "^4.17.21",
-    "@types/node": "^20.14.10",
+    "@types/node": "^20.14.13",
     "@types/nodemailer-sendgrid": "^1.0.3",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
@@ -71,8 +71,8 @@
     "eslint-plugin-import": "^2.29.1",
     "jest": "^29.7.0",
     "jest-config-commons-ui": "workspace:*",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "ts-node": "^10.9.2",
-    "typescript": "^5.5.3"
+    "typescript": "^5.5.4"
   }
 }
