+++ apps/uibook/package.json
@@ -30,41 +30,43 @@
   "dependencies": {
     "@commons-ui/core": "workspace:*",
     "@commons-ui/next": "workspace:*",
-    "@emotion/cache": "^11.11.0",
-    "@emotion/react": "^11.11.4",
+    "@emotion/cache": "^11.13.1",
+    "@emotion/react": "^11.13.0",
     "@emotion/server": "^11.11.0",
-    "@emotion/styled": "^11.11.5",
+    "@emotion/styled": "^11.13.0",
     "@hurumap/core": "workspace:*",
     "@hurumap/next": "workspace:*",
-    "@mui/material": "^5.16.1",
-    "@mui/utils": "^5.16.1",
+    "@mui/material": "^5.16.6",
+    "@mui/utils": "^5.16.6",
     "@next/env": "^14.2.5",
     "css-loader": "^7.1.2",
+    "leaflet": "^1.9.4",
     "next": "^14.2.5",
     "prop-types": "^15.8.1",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-leaflet": "^4.2.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
+    "@babel/core": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
-    "@storybook/addon-essentials": "^8.2.2",
-    "@storybook/addon-interactions": "^8.2.2",
-    "@storybook/addon-links": "^8.2.2",
-    "@storybook/blocks": "^8.2.2",
-    "@storybook/cli": "^8.2.2",
-    "@storybook/nextjs": "^8.2.2",
-    "@storybook/react": "^8.2.2",
-    "@storybook/test": "^8.2.2",
-    "@types/node": "^20.14.10",
+    "@storybook/addon-essentials": "^8.2.6",
+    "@storybook/addon-interactions": "^8.2.6",
+    "@storybook/addon-links": "^8.2.6",
+    "@storybook/blocks": "^8.2.6",
+    "@storybook/cli": "^8.2.6",
+    "@storybook/nextjs": "^8.2.6",
+    "@storybook/react": "^8.2.6",
+    "@storybook/test": "^8.2.6",
+    "@types/node": "^20.14.13",
     "@types/react": "^18.3.3",
     "eslint": "^8.57.0",
     "eslint-config-commons-ui": "workspace:*",
     "eslint-import-resolver-webpack": "^0.13.8",
     "eslint-plugin-import": "^2.29.1",
-    "prettier": "^3.3.2",
-    "storybook": "^8.2.2",
-    "typescript": "^5.5.3",
+    "prettier": "^3.3.3",
+    "storybook": "^8.2.6",
+    "typescript": "^5.5.4",
     "webpack": "^5.93.0"
   },
   "engines": {
