--- apps/uibook/stories/HURUmap/next/Map.stories.js
@@ -1,182 +1,0 @@
-import { Map } from "@hurumap/next";
-import React from "react";
-
-export default {
-  title: "@hurumap/next/Map",
-  argTypes: {
-    geography: {
-      control: {
-        type: "object",
-      },
-    },
-    geometries: {
-      control: {
-        type: "object",
-      },
-    },
-    tileLayers: {
-      control: {
-        type: "object",
-      },
-    },
-    zoom: {
-      control: {
-        type: "number",
-      },
-    },
-    center: {
-      control: {
-        type: "object",
-      },
-    },
-    preferredChildren: {
-      control: {
-        type: "object",
-      },
-    },
-  },
-};
-
-function Template({ ...args }) {
-  return <Map {...args} />;
-}
-
-export const Default = Template.bind({});
-
-Default.args = {
-  center: [0.3051933453207569, 37.908818734483155],
-  zoom: 6.25,
-  preferredChildren: { county: ["subcounty"], country: ["county"] },
-  geography: {
-    level: "county",
-  },
-  geometries: {
-    parents: [],
-    children: {},
-    boundary: {
-      type: "Feature",
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [37.0875, -1.249],
-              [37.1039, -1.2648],
-              [37.1015, -1.2742],
-              [37.0879, -1.2775],
-              [37.0823, -1.2847],
-              [37.0787, -1.2838],
-              [37.0735, -1.2954],
-              [37.076, -1.3024],
-              [37.0265, -1.2919],
-              [37.0027, -1.3007],
-              [36.9955, -1.2814],
-              [36.9743, -1.2896],
-              [36.9843, -1.3148],
-              [36.9669, -1.3532],
-              [36.9345, -1.3388],
-              [36.9073, -1.3634],
-              [36.9239, -1.3939],
-              [36.9369, -1.3997],
-              [36.9443, -1.4067],
-              [36.9475, -1.4219],
-              [36.9559, -1.4293],
-              [36.9617, -1.4387],
-              [36.9612, -1.4439],
-              [36.9555, -1.4376],
-              [36.9319, -1.4329],
-              [36.8929, -1.4157],
-              [36.8907, -1.4138],
-              [36.8943, -1.4111],
-              [36.8853, -1.4072],
-              [36.8711, -1.4071],
-              [36.8693, -1.4039],
-              [36.8497, -1.3961],
-              [36.8443, -1.3876],
-              [36.8299, -1.3876],
-              [36.8199, -1.3815],
-              [36.8091, -1.3881],
-              [36.7931, -1.3858],
-              [36.7725, -1.3885],
-              [36.7349, -1.3784],
-              [36.7027, -1.3522],
-              [36.6955, -1.353],
-              [36.6805, -1.3289],
-              [36.6755, -1.3303],
-              [36.6697, -1.3261],
-              [36.6735, -1.3244],
-              [36.6641, -1.3197],
-              [36.6709, -1.306],
-              [36.6743, -1.3073],
-              [36.6817, -1.298],
-              [36.6891, -1.2827],
-              [36.6889, -1.2755],
-              [36.6919, -1.2755],
-              [36.6919, -1.2716],
-              [36.6969, -1.2675],
-              [36.6941, -1.2639],
-              [36.6967, -1.2606],
-              [36.7015, -1.2617],
-              [36.7043, -1.2586],
-              [36.7197, -1.2637],
-              [36.7311, -1.2348],
-              [36.7459, -1.2371],
-              [36.7566, -1.2204],
-              [36.7789, -1.2268],
-              [36.7889, -1.2078],
-              [36.7852, -1.2048],
-              [36.7915, -1.1918],
-              [36.8351, -1.2161],
-              [36.8333, -1.2091],
-              [36.8501, -1.2131],
-              [36.8527, -1.2026],
-              [36.8601, -1.2019],
-              [36.8639, -1.1904],
-              [36.8785, -1.1964],
-              [36.8969, -1.1923],
-              [36.8891, -1.1743],
-              [36.9022, -1.1606],
-              [36.9203, -1.1632],
-              [36.9415, -1.1746],
-              [36.9135, -1.208],
-              [36.9317, -1.2233],
-              [36.9391, -1.2222],
-              [36.9585, -1.2288],
-              [36.9887, -1.2241],
-              [37.0045, -1.2366],
-              [37.0125, -1.2355],
-              [37.0165, -1.2285],
-              [37.0383, -1.2132],
-              [37.0611, -1.2069],
-              [37.0809, -1.2373],
-              [37.0877, -1.2341],
-              [37.0915, -1.2379],
-              [37.0955, -1.2448],
-              [37.0875, -1.249],
-            ],
-          ],
-        ],
-      },
-      properties: {
-        code: "47",
-        name: "NAIROBI",
-        area: 0.057422949,
-        parent: "KE",
-        level: "county",
-        version: "KE Boundaries",
-      },
-    },
-  },
-  tileLayers: [
-    {
-      url: "https://{s}.basemaps.cartocdn.com/rastertiles/voyager_nolabels/{z}/{x}/{y}.png",
-      pane: "tiles",
-      zIndex: 200,
-    },
-    {
-      url: "https://{s}.basemaps.cartocdn.com/rastertiles/voyager_only_labels/{z}/{x}/{y}.png",
-      pane: "labels",
-      zIndex: 650,
-    },
-  ],
-};
