+++ apps/uibook/stories/hurumap/core/LocationHighlight.stories.js
