+++ apps/vpnmanager/package.json
@@ -16,15 +16,15 @@
     "@babel/preset-react": "^7.24.7",
     "@commons-ui/core": "workspace:*",
     "@commons-ui/next": "workspace:*",
-    "@emotion/cache": "^11.11.0",
-    "@emotion/react": "^11.11.4",
+    "@emotion/cache": "^11.13.1",
+    "@emotion/react": "^11.13.0",
     "@emotion/server": "^11.11.0",
-    "@emotion/styled": "^11.11.5",
-    "@mui/material": "^5.16.1",
-    "@mui/utils": "^5.16.1",
+    "@emotion/styled": "^11.13.0",
+    "@mui/material": "^5.16.6",
+    "@mui/utils": "^5.16.6",
     "@next/env": "^14.2.5",
     "@sendgrid/mail": "^8.1.3",
-    "@sentry/nextjs": "^8.17.0",
+    "@sentry/nextjs": "^8.21.0",
     "@svgr/webpack": "^8.1.0",
     "@types/jest": "^29.5.12",
     "googleapis": "^133.0.0",
@@ -35,14 +35,14 @@
     "tsc-alias": "^1.8.10"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
+    "@babel/core": "^7.25.2",
     "@commons-ui/testing-library": "workspace:*",
-    "@types/node": "^20.14.10",
+    "@types/node": "^20.14.13",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "eslint": "^8.57.0",
     "eslint-config-next": "^14.2.5",
     "eslint-config-prettier": "^9.1.0",
-    "typescript": "^5.5.3"
+    "typescript": "^5.5.4"
   }
 }
