+++ package.json
@@ -28,14 +28,14 @@
     "@actions/exec": "^1.1.1",
     "@changesets/changelog-github": "^0.5.0",
     "@changesets/cli": "^2.27.7",
-    "@playwright/test": "^1.45.1",
+    "@playwright/test": "^1.45.3",
     "eslint": "^8.57.0",
     "eslint-config-commons-ui": "workspace:*",
     "husky": "^8.0.3",
     "lint-staged": "^15.2.7",
     "playwright-config-commons-ui": "workspace:*",
-    "prettier": "^3.3.2",
-    "turbo": "^2.0.6"
+    "prettier": "^3.3.3",
+    "turbo": "^2.0.10"
   },
   "engines": {
     "node": "20.x",
