+++ packages/commons-ui-core/package.json
@@ -28,13 +28,13 @@
     "clean": "rm -rf .turbo node_modules dist"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
+    "@babel/core": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
     "@commons-ui/testing-library": "workspace:*",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
-    "@mui/material": "^5.16.1",
-    "@mui/utils": "^5.16.1",
+    "@emotion/react": "^11.13.0",
+    "@emotion/styled": "^11.13.0",
+    "@mui/material": "^5.16.6",
+    "@mui/utils": "^5.16.6",
     "@types/react": "^18.3.3",
     "babel-loader": "^9.1.3",
     "eslint": "^8.57.0",
@@ -42,12 +42,12 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-config-commons-ui": "workspace:*",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-test-renderer": "^18.3.1",
     "require-from-string": "^2.0.2",
-    "typescript": "^5.5.3",
+    "typescript": "^5.5.4",
     "webpack": "^5.93.0"
   },
   "peerDependencies": {
