+++ packages/commons-ui-next/package.json
@@ -30,11 +30,11 @@
     "clean": "rm -rf .turbo node_modules dist"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
+    "@babel/core": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
     "@commons-ui/core": "workspace:*",
     "@commons-ui/testing-library": "workspace:*",
-    "@mui/material": "^5.16.1",
+    "@mui/material": "^5.16.6",
     "babel-loader": "^9.1.3",
     "eslint": "^8.57.0",
     "eslint-config-commons-ui": "workspace:*",
@@ -45,7 +45,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "require-from-string": "^2.0.2",
-    "typescript": "^5.5.3",
+    "typescript": "^5.5.4",
     "webpack": "^5.93.0"
   },
   "peerDependencies": {
