+++ packages/commons-ui-testing-library/package.json
@@ -32,11 +32,11 @@
     "clean": "rm -rf .turbo node_modules dist"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
+    "@babel/core": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
-    "@mui/material": "^5.16.1",
+    "@emotion/react": "^11.13.0",
+    "@emotion/styled": "^11.13.0",
+    "@mui/material": "^5.16.6",
     "@types/react": "^18.3.3",
     "babel-loader": "^9.1.3",
     "eslint": "^8.57.0",
@@ -44,11 +44,11 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-config-commons-ui": "workspace:*",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-test-renderer": "^18.3.1",
-    "typescript": "^5.5.3",
+    "typescript": "^5.5.4",
     "webpack": "^5.93.0"
   },
   "peerDependencies": {
@@ -60,7 +60,7 @@
     "react-dom": "^18.2.0"
   },
   "dependencies": {
-    "@testing-library/jest-dom": "^6.4.6",
+    "@testing-library/jest-dom": "^6.4.8",
     "@testing-library/react": "^14.3.1",
     "prop-types": "^15.8.1"
   }
