+++ packages/eslint-config-commons-ui/package.json
@@ -13,12 +13,12 @@
     "clean": "rm -rf .turbo node_modules"
   },
   "dependencies": {
-    "@babel/eslint-parser": "^7.24.8",
+    "@babel/eslint-parser": "^7.25.1",
     "@babel/preset-react": "^7.24.7",
     "@next/eslint-plugin-next": "^14.2.5",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-config-turbo": "^2.0.6",
+    "eslint-config-turbo": "^2.0.10",
     "eslint-import-resolver-jsconfig": "^1.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jest": "^27.9.0",
@@ -27,17 +27,17 @@
     "eslint-plugin-jsx-a11y": "^6.9.0",
     "eslint-plugin-markdown": "^3.0.1",
     "eslint-plugin-playwright": "^0.19.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-react": "^7.34.3",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.35.0",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-testing-library": "^6.2.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
+    "@babel/core": "^7.25.2",
     "eslint": "^8.57.0",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "react": "^18.3.1",
-    "typescript": "^5.5.3"
+    "typescript": "^5.5.4"
   },
   "peerDependencies": {
     "eslint": "^8.11.0",
