+++ packages/hurumap-core/package.json
@@ -31,13 +31,13 @@
     "clean": "rm -rf .turbo node_modules dist"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
+    "@babel/core": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
     "@commons-ui/testing-library": "workspace:*",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
-    "@mui/material": "^5.16.1",
-    "@mui/utils": "^5.16.1",
+    "@emotion/react": "^11.13.0",
+    "@emotion/styled": "^11.13.0",
+    "@mui/material": "^5.16.6",
+    "@mui/utils": "^5.16.6",
     "@types/react": "^18.3.3",
     "babel-loader": "^9.1.3",
     "eslint": "^8.57.0",
@@ -46,13 +46,13 @@
     "jest": "^29.7.0",
     "jest-config-commons-ui": "workspace:*",
     "leaflet": "^1.9.4",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-leaflet": "^4.2.1",
     "react-test-renderer": "^18.3.1",
     "require-from-string": "^2.0.2",
-    "typescript": "^5.5.3",
+    "typescript": "^5.5.4",
     "webpack": "^5.93.0"
   },
   "peerDependencies": {
@@ -61,10 +61,8 @@
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.14.0",
     "@mui/utils": "^5.12.3",
-    "leaflet": "^1.9.4",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-leaflet": "^4.2.1"
+    "react-dom": "^18.2.0"
   },
   "dependencies": {
     "prop-types": "^15.8.1"
