+++ packages/hurumap-next/package.json
@@ -31,38 +31,43 @@
     "clean": "rm -rf .turbo node_modules dist"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
+    "@babel/core": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
     "@commons-ui/core": "workspace:*",
+    "@commons-ui/next": "workspace:*",
     "@commons-ui/testing-library": "workspace:*",
+    "@emotion/react": "^11.13.0",
+    "@emotion/styled": "^11.13.0",
     "@hurumap/core": "workspace:*",
-    "@mui/material": "^5.16.1",
+    "@mui/material": "^5.16.6",
     "babel-loader": "^9.1.3",
     "eslint": "^8.57.0",
     "eslint-config-commons-ui": "workspace:*",
     "jest": "^29.7.0",
     "jest-config-commons-ui": "workspace:*",
+    "leaflet": "^1.9.4",
     "next": "^14.2.5",
     "prop-types": "^15.8.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-leaflet": "^4.2.1",
     "require-from-string": "^2.0.2",
-    "typescript": "^5.5.3",
+    "typescript": "^5.5.4",
     "webpack": "^5.93.0"
   },
   "peerDependencies": {
     "@babel/core": "^7.22.9",
     "@commons-ui/core": "workspace:*",
+    "@commons-ui/next": "workspace:*",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
     "@hurumap/core": "workspace:*",
     "@mui/material": "^5.14.0",
-    "next": "~13.4.11",
+    "leaflet": "^1.9.4",
+    "next": "^14.2.5",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "dependencies": {
-    "@commons-ui/next": "workspace:*",
-    "leaflet": "^1.9.4",
+    "react-dom": "^18.2.0",
     "react-leaflet": "^4.2.1"
   }
 }
