+++ packages/hurumap-next/src/Map/Layers.js
@@ -216,7 +216,6 @@ function Layers({
     siblingRef,
   ]);
 
-  console.log("BOOM: ", { selectedBoundary });
   return (
     <>
       <FeatureGroup ref={siblingRef}>
