+++ packages/playwright-config-commons-ui/package.json
@@ -12,15 +12,15 @@
     "clean": "rm -rf .turbo node_modules"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.8",
-    "@babel/preset-env": "^7.24.8",
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.25.2",
     "@babel/preset-react": "^7.24.7",
-    "@playwright/test": "^1.45.1",
+    "@playwright/test": "^1.45.3",
     "eslint": "^8.57.0",
     "eslint-config-commons-ui": "workspace:*",
-    "prettier": "^3.3.2",
+    "prettier": "^3.3.3",
     "react": "^18.3.1",
-    "typescript": "^5.5.3"
+    "typescript": "^5.5.4"
   },
   "peerDependencies": {
     "@playwright/test": "^1.19.2"
